refactoring expression table
